creat addRank for registring a member in cometition

+++ src/main/java/com/redone/aftas/repositories/HuntingRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface HuntingRepository extends JpaRepository<Hunting ,Long> {
+public interface HuntingRepository extends JpaRepository<Hunting ,Integer> {
 }
